[#ic-87] [#ic-87] refactor: all event defintions on the same file
